+++ src/interfaces/cursor_paginated_result.ts
@@ -1,0 +1,7 @@
+import { PaginatedResult } from "./paginated_result.js";
+
+export interface CursorPaginatedResult<T = any> extends PaginatedResult<T> {
+  readonly nextCursor: string | null;
+
+  hasNextCursor(): boolean;
+}