+++ src/types/common_get_params.ts
@@ -12,3 +12,9 @@ export type PaginationParams = {
 };
 
 export type ProjectWithPagination = ProjectOnly & PaginationParams;
+
+export type CursorPagination = {
+  limit?: number;
+  pagination?: "offset" | "cursor";
+  cursor?: string;
+};