+++ src/types/keys.ts
@@ -6,6 +6,7 @@ import type { TranslationData } from "./translations.js";
 import type {
   ProjectWithPagination,
   ProjectOnly,
+  CursorPagination,
 } from "./common_get_params.js";
 import type { NumericBool } from "./numeric_bool.js";
 
@@ -61,21 +62,22 @@ export type KeysBulkDeleted = {
   branch?: string;
 };
 
-export type KeyParamsWithPagination = ProjectWithPagination & {
-  disable_references?: NumericBool;
-  include_comments?: NumericBool;
-  include_screenshots?: NumericBool;
-  include_translations?: NumericBool;
-  filter_translation_lang_ids?: string;
-  filter_tags?: string;
-  filter_filenames?: string;
-  filter_keys?: string;
-  filter_key_ids?: string;
-  filter_platforms?: string;
-  filter_untranslated?: NumericBool;
-  filter_qa_issues?: string;
-  filter_archived?: string;
-};
+export type KeyParamsWithPagination = ProjectWithPagination &
+  CursorPagination & {
+    disable_references?: NumericBool;
+    include_comments?: NumericBool;
+    include_screenshots?: NumericBool;
+    include_translations?: NumericBool;
+    filter_translation_lang_ids?: string;
+    filter_tags?: string;
+    filter_filenames?: string;
+    filter_keys?: string;
+    filter_key_ids?: string;
+    filter_platforms?: string;
+    filter_untranslated?: NumericBool;
+    filter_qa_issues?: string;
+    filter_archived?: string;
+  };
 
 export type GetKeyParams = ProjectOnly & {
   disable_references?: NumericBool;